Delete properly, and parse full source
git-svn-id: http://svn.dojotoolkit.org/src@6362 560b804f-0ae3-0310-86f3-f6aa0a117693

+++ trunk/docscripts/inc/Plugins.php
@@ -67,8 +67,10 @@ class Plugins
       foreach($files as $file){ 
         if($file{0} != '.'){ 
           if(is_dir($directory . '/' . $file)){ 
-            deltree($directory, $file);
-            rmdir($directory . '/' . $file);
+            $this->delTree($directory, $file);
+						if (!in_array('.svn',  $files) && scandir($directory . '/' . $file) == 2) {
+							rmdir($directory . '/' . $file);
+						}
           }else{ 
             unlink($directory . '/' . $file); 
           }
